Second Arcade Session
I added my unattend.xml File into my .iso Image with a Tool called AnyBurn. I tried out a bunch of tools for writing into ISO Files but this one was clearly the best. See my new File to find the link to the ISO. It was sadly to large to upload into the repository directly. Im now ready to move on and to try out my unattended installation.